+++ samples/sampler/stories/qa/Header.js
@@ -2,6 +2,7 @@ import {boolean, text, select} from '@enact/storybook-utils/addons/knobs';
 import {mergeComponentMetadata} from '@enact/storybook-utils';
 import React from 'react';
 import {storiesOf} from '@storybook/react';
+import clone from 'ramda/src/clone';
 
 import Button from '@enact/sandstone/Button';
 import Input from '@enact/sandstone/Input';
@@ -29,16 +30,19 @@ const prop = {
 		none: null,
 		steps: <Steps current={3} total={5} />
 	},
+	aboveSelection: ['none', 'steps'],
 	buttons: {
 		'no buttons': null,
-		'1 button': <Button icon="ellipsis" />,
-		'2 buttons': <React.Fragment>
-			<Button icon="search" />
-			<Button icon="ellipsis" />
-		</React.Fragment>
+		'1 button': <Button key="button-ellipsis1" size="small" icon="ellipsis" />,
+		'2 buttons': [
+			<Button key="button-search2" size="small" icon="search" />,
+			<Button key="button-ellipsis2" size="small" icon="ellipsis" />
+		]
 	},
 	buttonsSelection: ['no buttons', '1 button', '2 buttons'],
-	marqueeOn: ['hover', 'render']
+	backgroundOpacity: ['opaque', 'transparent'],
+	marqueeOn: ['hover', 'render'],
+	type: ['standard', 'compact', 'wizard', 'mini']
 };
 
 const headerStoryConfig = {
@@ -47,238 +51,244 @@ const headerStoryConfig = {
 	}
 };
 
-function headerComponents () {
-	const slotAboveSelection = select('slotAbove', ['none', 'steps'], Config);
-	const slotAbove = prop.above[slotAboveSelection];
-	const slotBeforeSelection = select('slotBefore', prop.buttonsSelection, Config);
-	const slotBefore = prop.buttons[slotBeforeSelection];
-	const slotAfterSelection = select('slotAfter', prop.buttonsSelection, Config);
-	const slotAfter = prop.buttons[slotAfterSelection];
-	const childrenSelection = select('children', prop.buttonsSelection, Config);
-	const children = prop.buttons[childrenSelection];
-
-	return {slotAbove, slotBefore, slotAfter, children};
-}
+const commonProps = (customDefaults) => {
+	const customizedConfig = Object.assign({}, Config); // Shallow copy this fn into a normal object
+	customizedConfig.defaultProps = Object.assign(
+		{}, // Fresh new defaltProps object
+		clone(Config.defaultProps), // Deep copy the defaultProps from our object into a fresh defaultProps object of our shallow copy (preserving all other props as their original references)
+		{
+			backButtonBackgroundOpacity: 'opaque',
+			centered: false,
+			closeButtonBackgroundOpacity: 'opaque',
+			noBackButton: false,
+			noCloseButton: false,
+			slotAbove: 'none',
+			slotBefore: 'no buttons',
+			slotAfter: 'no buttons',
+			children: 'no buttons'
+		}, // Story global defaults (things not represented by defaultProps on the real component)
+		customDefaults // Individual story defaults, preferences
+	);
+	return {
+		type: select('type', prop.type, customizedConfig),
+		centered: boolean('centered', customizedConfig),
+		backButtonAvailable: boolean('backButtonAvailable', customizedConfig),
+		backButtonBackgroundOpacity: select('backButtonBackgroundOpacity', prop.backgroundOpacity, customizedConfig),
+		closeButtonBackgroundOpacity: select('closeButtonBackgroundOpacity', prop.backgroundOpacity, customizedConfig),
+		noBackButton: boolean('noBackButton', customizedConfig),
+		noCloseButton: boolean('noCloseButton', customizedConfig),
+		marqueeOn: select('marqueeOn', prop.marqueeOn, customizedConfig),
+		slotAbove: prop.above[select('slotAbove', prop.aboveSelection, customizedConfig)],
+		slotBefore: prop.buttons[select('slotBefore', prop.buttonsSelection, customizedConfig)],
+		slotAfter: prop.buttons[select('slotAfter', prop.buttonsSelection, customizedConfig)],
+		children: prop.buttons[select('children', prop.buttonsSelection, customizedConfig)]
+	};
+};
 
 storiesOf('Header', module)
 	.add(
 		'just title',
+		// The Fragment (or any node, really; could be a <div> instead) is actually needed by
+		// Storybook to properly apply changes from the knobs to the stories' children that occupy
+		// the outermost node. This is most visible when the `noHeader` prop is given and several
+		// (not all) of the knobs fail to apply.
 		() => {
-			const {slotAbove, slotBefore, slotAfter, children} = headerComponents();
-			return (
+			return (<React.Fragment>
 				<Header
 					title={text('title', Config, inputData.shortTitle)}
-					marqueeOn={select('marqueeOn', prop.marqueeOn, Config)}
-					slotAbove={slotAbove}
-					slotAfter={slotAfter}
-					slotBefore={slotBefore}
-				>
-					{children}
-				</Header>
-			);
-		}, headerStoryConfig
-	)
-	.add(
-		'just title, Compact',
-		() => {
-			const {slotAbove, slotBefore, slotAfter, children} = headerComponents();
-			return (
-				<Header
-					type="compact"
-					title={text('title', Config, inputData.shortTitle)}
-					marqueeOn={select('marqueeOn', prop.marqueeOn, Config)}
-					slotAbove={slotAbove}
-					slotAfter={slotAfter}
-					slotBefore={slotBefore}
-				>
-					{children}
-				</Header>
-			);
+					{...commonProps()}
+				/>
+			</React.Fragment>);
 		}, headerStoryConfig
 	)
 	.add(
 		'short titles',
 		() => {
-			const {slotAbove, slotBefore, slotAfter, children} = headerComponents();
-			return (
+			return (<React.Fragment>
 				<Header
 					title={text('title', Config, inputData.shortTitle)}
 					subtitle={text('subtitle', Config, inputData.shortSubtitle)}
-					marqueeOn={select('marqueeOn', prop.marqueeOn, Config)}
-					slotAbove={slotAbove}
-					slotAfter={slotAfter}
-					slotBefore={slotBefore}
-				>
-					{children}
-				</Header>
-			);
-		}, headerStoryConfig
-	)
-	.add(
-		'short titles, Compact',
-		() => {
-			const {slotAbove, slotBefore, slotAfter, children} = headerComponents();
-			return (
-				<Header
-					type="compact"
-					title={text('title', Config, inputData.shortTitle)}
-					subtitle={text('subtitle', Config, inputData.shortSubtitle)}
-					marqueeOn={select('marqueeOn', prop.marqueeOn, Config)}
-					slotAbove={slotAbove}
-					slotAfter={slotAfter}
-					slotBefore={slotBefore}
-				>
-					{children}
-				</Header>
-			);
+					{...commonProps()}
+				/>
+			</React.Fragment>);
 		}, headerStoryConfig
 	)
 	.add(
 		'long titles',
 		() => {
-			const {slotAbove, slotBefore, slotAfter, children} = headerComponents();
-			return (
+			return (<React.Fragment>
 				<Header
 					title={text('title', Config, inputData.longTitle)}
 					subtitle={text('subtitle', Config, inputData.longSubtitle)}
-					marqueeOn={select('marqueeOn', prop.marqueeOn, Config)}
-					slotAbove={slotAbove}
-					slotAfter={slotAfter}
-					slotBefore={slotBefore}
-				>
-					{children}
-				</Header>
-			);
+					{...commonProps()}
+				/>
+			</React.Fragment>);
 		}, headerStoryConfig
-	)
-	.add(
-		'long titles, Compact',
-		() => {
-			const {slotAbove, slotBefore, slotAfter, children} = headerComponents();
-			return (
-				<Header
-					type="compact"
-					title={text('title', Config, inputData.longTitle)}
-					subtitle={text('subtitle', Config, inputData.longSubtitle)}
-					marqueeOn={select('marqueeOn', prop.marqueeOn, Config)}
-					slotAbove={slotAbove}
-					slotAfter={slotAfter}
-					slotBefore={slotBefore}
-				>
-					{children}
-				</Header>
-			);
-		}, headerStoryConfig
-	)
+	);
+
+storiesOf('Header.Locale', module)
 	.add(
 		'RTL text',
 		() => {
-			const {slotAbove, slotBefore, slotAfter, children} = headerComponents();
-			return (
+			return (<React.Fragment>
 				<Header
 					title={text('title', Config, inputData.shortRtlTitle)}
 					subtitle={text('subtitle', Config, inputData.shortRtlSubtitle)}
-					marqueeOn={select('marqueeOn', prop.marqueeOn, Config)}
-					slotAbove={slotAbove}
-					slotAfter={slotAfter}
-					slotBefore={slotBefore}
-				>
-					{children}
-				</Header>
-			);
-		}, headerStoryConfig
-	)
-	.add(
-		'RTL text, Compact',
-		() => {
-			const {slotAbove, slotBefore, slotAfter, children} = headerComponents();
-			return (
-				<Header
-					type="compact"
-					title={text('title', Config, inputData.shortRtlTitle)}
-					subtitle={text('subtitle', Config, inputData.shortRtlSubtitle)}
-					marqueeOn={select('marqueeOn', prop.marqueeOn, Config)}
-					slotAbove={slotAbove}
-					slotAfter={slotAfter}
-					slotBefore={slotBefore}
-				>
-					{children}
-				</Header>
-			);
+					{...commonProps()}
+				/>
+			</React.Fragment>);
 		}, headerStoryConfig
 	)
 	.add(
 		'RTL text, long title',
 		() => {
-			const {slotAbove, slotBefore, slotAfter, children} = headerComponents();
-			return (
+			return (<React.Fragment>
 				<Header
 					title={text('title', Config, inputData. longRtlTitle)}
 					subtitle={text('subtitle', Config, inputData.shortRtlSubtitle)}
-					marqueeOn={select('marqueeOn', prop.marqueeOn, Config)}
-					slotAbove={slotAbove}
-					slotAfter={slotAfter}
-					slotBefore={slotBefore}
-				>
-					{children}
-				</Header>
-			);
-		}, headerStoryConfig
-	)
-	.add(
-		'RTL text, long title, Compact',
-		() => {
-			const {slotAbove, slotBefore, slotAfter, children} = headerComponents();
-			return (
-				<Header
-					type="compact"
-					title={text('title', Config, inputData. longRtlTitle)}
-					subtitle={text('subtitle', Config, inputData.shortRtlSubtitle)}
-					marqueeOn={select('marqueeOn', prop.marqueeOn, Config)}
-					slotAbove={slotAbove}
-					slotAfter={slotAfter}
-					slotBefore={slotBefore}
-				>
-					{children}
-				</Header>
-			);
+					{...commonProps()}
+				/>
+			</React.Fragment>);
 		}, headerStoryConfig
 	)
 	.add(
 		'tall-glyphs',
 		() => {
-			const {slotAbove, slotBefore, slotAfter, children} = headerComponents();
-			return (
+			return (<React.Fragment>
 				<Header
 					title={text('title', Config, inputData.tallText)}
 					subtitle={text('subtitle', Config, inputData.tallText)}
-					marqueeOn={select('marqueeOn', prop.marqueeOn, Config)}
-					slotAbove={slotAbove}
-					slotAfter={slotAfter}
-					slotBefore={slotBefore}
-				>
-					{children}
-				</Header>
-			);
+					{...commonProps()}
+				/>
+			</React.Fragment>);
+		}, headerStoryConfig
+	);
+
+const wizardDefaultProps = {
+	type: 'wizard',
+	centered: true,
+	noCloseButton: true,
+	slotAbove: 'steps',
+	slotBefore: '1 button',
+	slotAfter: '1 button'
+};
+storiesOf('Header.Wizard', module)
+	.add(
+		'just title',
+		() => {
+			return (<React.Fragment>
+				<Header
+					title={text('title', Config, inputData.shortTitle)}
+					{...commonProps(wizardDefaultProps)}
+				/>
+			</React.Fragment>);
 		}, headerStoryConfig
 	)
 	.add(
-		'tall-glyphs, Compact',
+		'short titles',
 		() => {
-			const {slotAbove, slotBefore, slotAfter, children} = headerComponents();
-			return (
+			return (<React.Fragment>
 				<Header
-					type="compact"
-					title={text('title', Config, inputData.tallText)}
-					subtitle={text('subtitle', Config, inputData.tallText)}
-					marqueeOn={select('marqueeOn', prop.marqueeOn, Config)}
-					slotAbove={slotAbove}
-					slotAfter={slotAfter}
-					slotBefore={slotBefore}
-				>
-					{children}
-				</Header>
-			);
+					title={text('title', Config, inputData.shortTitle)}
+					subtitle={text('subtitle', Config, inputData.shortSubtitle)}
+					{...commonProps(wizardDefaultProps)}
+				/>
+			</React.Fragment>);
+		}, headerStoryConfig
+	)
+	.add(
+		'long titles',
+		() => {
+			return (<React.Fragment>
+				<Header
+					title={text('title', Config, inputData.longTitle)}
+					subtitle={text('subtitle', Config, inputData.longSubtitle)}
+					{...commonProps(wizardDefaultProps)}
+				/>
+			</React.Fragment>);
+		}, headerStoryConfig
+	);
+
+
+const compactDefaultProps = {
+	type: 'compact'
+};
+storiesOf('Header.Compact', module)
+	.add(
+		'just title',
+		() => {
+			return (<React.Fragment>
+				<Header
+					title={text('title', Config, inputData.shortTitle)}
+					{...commonProps(compactDefaultProps)}
+				/>
+			</React.Fragment>);
+		}, headerStoryConfig
+	)
+	.add(
+		'short titles',
+		() => {
+			return (<React.Fragment>
+				<Header
+					title={text('title', Config, inputData.shortTitle)}
+					subtitle={text('subtitle', Config, inputData.shortSubtitle)}
+					{...commonProps(compactDefaultProps)}
+				/>
+			</React.Fragment>);
+		}, headerStoryConfig
+	)
+	.add(
+		'long titles',
+		() => {
+			return (<React.Fragment>
+				<Header
+					title={text('title', Config, inputData.longTitle)}
+					subtitle={text('subtitle', Config, inputData.longSubtitle)}
+					{...commonProps(compactDefaultProps)}
+				/>
+			</React.Fragment>);
+		}, headerStoryConfig
+	);
+
+
+const miniDefaultProps = {
+	type: 'mini',
+	noCloseButton: true
+};
+storiesOf('Header.Mini', module)
+	.add(
+		'just title',
+		() => {
+			return (<React.Fragment>
+				<Header
+					title={text('title', Config, inputData.shortTitle)}
+					{...commonProps(miniDefaultProps)}
+				/>
+			</React.Fragment>);
+		}, headerStoryConfig
+	)
+	.add(
+		'short titles',
+		() => {
+			return (<React.Fragment>
+				<Header
+					title={text('title', Config, inputData.shortTitle)}
+					subtitle={text('subtitle', Config, inputData.shortSubtitle)}
+					{...commonProps(miniDefaultProps)}
+				/>
+			</React.Fragment>);
+		}, headerStoryConfig
+	)
+	.add(
+		'long titles',
+		() => {
+			return (<React.Fragment>
+				<Header
+					title={text('title', Config, inputData.longTitle)}
+					subtitle={text('subtitle', Config, inputData.longSubtitle)}
+					{...commonProps(miniDefaultProps)}
+				/>
+			</React.Fragment>);
 		}, headerStoryConfig
 	);
 
@@ -286,44 +296,32 @@ storiesOf('Header.Input', module)
 	.add(
 		'tall-glyphs',
 		() => {
-			const {slotAbove, slotBefore, slotAfter, children} = headerComponents();
 			const input = boolean('Input Mode', Config, true) ? <Input placeholder={text('placeholder', Config, inputData.longTitle)} dismissOnEnter={boolean('Input dismissOnEnter', Config, true)} /> : null;
 			const showInput = boolean('showInput', Config, true);
-			return (
+			return (<React.Fragment>
 				<Header
 					title={text('title', Config, inputData.tallText)}
 					headerInput={input}
 					subtitle={text('subtitle', Config, inputData.longSubtitle)}
-					marqueeOn={select('marqueeOn', prop.marqueeOn, Config)}
 					showInput={showInput}
-					slotAbove={slotAbove}
-					slotAfter={slotAfter}
-					slotBefore={slotBefore}
-				>
-					{children}
-				</Header>
-			);
+					{...commonProps()}
+				/>
+			</React.Fragment>);
 		}, headerStoryConfig
 	)
 	.add(
 		'long text',
 		() => {
-			const {slotAbove, slotBefore, slotAfter, children} = headerComponents();
 			const input = boolean('Input Mode', Config, true) ? <Input placeholder={text('placeholder', Config, inputData.longTitle)} dismissOnEnter={boolean('Input dismissOnEnter', Config, true)} /> : null;
 			const showInput = boolean('showInput', Config, true);
-			return (
+			return (<React.Fragment>
 				<Header
 					headerInput={input}
 					title={text('title', Config, inputData.longTitle)}
 					subtitle={text('subtitle', Config, inputData.longSubtitle)}
-					marqueeOn={select('marqueeOn', prop.marqueeOn, Config)}
 					showInput={showInput}
-					slotAbove={slotAbove}
-					slotAfter={slotAfter}
-					slotBefore={slotBefore}
-				>
-					{children}
-				</Header>
-			);
+					{...commonProps()}
+				/>
+			</React.Fragment>);
 		}, headerStoryConfig
 	);